backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
rank-bm25==0.2.2
PyMuPDF==1.24.14
python-docx==1.1.2
python-pptx==1.0.2
google-api-python-client==2.143.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
google-generativeai>=0.8.6
python-dotenv==1.0.1
httpx==0.27.0
pytest==8.3.3
